ai/requirements_app.txt
# Requirements for Streamlit app
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0
